chore: move internal/suspiciousnames to cmd/frontend/internal
Just another package that is now only used by frontend, so moving it around to make clearer which service needs this package.

Test plan: Go compiler doesn't complain about moved directory.
