+++ cmd/frontend/graphqlbackend/BUILD.bazel
@@ -254,6 +254,7 @@ go_library(
         "//cmd/frontend/internal/processrestart",
         "//cmd/frontend/internal/search/logs",
         "//cmd/frontend/internal/ssc",
+        "//cmd/frontend/internal/suspiciousnames",
         "//internal/actor",
         "//internal/adminanalytics",
         "//internal/api",
@@ -331,7 +332,6 @@ go_library(
         "//internal/settings",
         "//internal/siteid",
         "//internal/src-prometheus",
-        "//internal/suspiciousnames",
         "//internal/symbols",
         "//internal/temporarysettings",
         "//internal/trace",
