+++ cmd/frontend/internal/auth/userpasswd/BUILD.bazel
@@ -26,6 +26,7 @@ go_library(
         "//cmd/frontend/hubspot/hubspotutil",
         "//cmd/frontend/internal/auth/providers",
         "//cmd/frontend/internal/auth/session",
+        "//cmd/frontend/internal/suspiciousnames",
         "//internal/actor",
         "//internal/auth",
         "//internal/authz",
@@ -40,7 +41,6 @@ go_library(
         "//internal/lazyregexp",
         "//internal/rcache",
         "//internal/redispool",
-        "//internal/suspiciousnames",
         "//internal/telemetry",
         "//internal/telemetry/telemetrystore/teestore",
         "//internal/txemail",
