+++ cmd/frontend/internal/suspiciousnames/BUILD.bazel
@@ -3,9 +3,9 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "suspiciousnames",
     srcs = ["names.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/suspiciousnames",
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/suspiciousnames",
     tags = [TAG_PLATFORM_SOURCE],
-    visibility = ["//:__subpackages__"],
+    visibility = ["//cmd/frontend:__subpackages__"],
     deps = [
         "//internal/dotcom",
         "//lib/errors",
